different OS vars
Former-commit-id: abb0bafd8bcddbfcd7f4ce48e3943488be91c629 [formerly 9eb1d874ca2de57662b43c8036ab079643e4309e]
Former-commit-id: a1b560e729b6be67fa8f0841909f9115def2a791

+++ SConstruct
@@ -41,14 +41,24 @@ import subprocess
 import SCons.Script
 import SCons.SConf
 
-
+MACOSX = (platform.system() == 'Darwin')
+WINDOWS = (platform.system() == 'Windows')
+LINUX = (platform.system() == 'Linux')
 
 # URL where we have most of our tgz files for libraries, modules and packages.
 URL_BASE = os.environ['SCIPION_URL_SOFTWARE']
 
-# Define our builders.
-download = Builder(action='wget -nv $SOURCE -c -O $TARGET')
-untar = Builder(action='tar -C $cdir --recursive-unlink -xzf $SOURCE')
+# Define our builders
+if LINUX:
+    download = Builder(action='wget -nv --show-progress -c -O $TARGET $SOURCE')
+    untar = Builder(action='tar -C $cdir --recursive-unlink -xzf $SOURCE')
+elif MACOSX:
+    download = Builder(action='curl -L "$SOURCE" -o "$TARGET"')
+    untar = Builder(action='tar -C $cdir -xzf $SOURCE')
+else:
+    print 'OS not tested yet'
+    Exit(1)
+
 
 # Create the environment the whole build will use.
 env = Environment(ENV=os.environ,
